+++ app/Http/Requests/Auth/PasswordConfirmRequest.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Requests\Auth;
 
-use App\Http\Requests\FormRequest;
+use Enraiged\Support\Requests\FormRequest;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\ValidationException;