+++ app/Http/Requests/Auth/PasswordResetRequest.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Requests\Auth;
 
-use App\Http\Requests\FormRequest;
+use Enraiged\Support\Requests\FormRequest;
 use Illuminate\Auth\Events\PasswordReset;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Password;