+++ packages/enraiged/src/Accounts/Traits/Tables/Actions.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Enraiged\Accounts\Traits\Tables;
+
+trait Actions
+{
+    /** @var  array  The table actions. */
+    protected $actions = [
+        'show', 'edit', 'destroy', // 'export',
+    ];
+}